+++ images/base/alpine-3.14/CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 3.14-alpine
 
 - base image, based on Alpine 3.14-alpine
+- add removing /var/cache as part cleaning